Move instrumentation implementation to trace module
